Verify the TLS cert when making POST request
In 3932b05 the verify=False flag was added to the requests.post() call
to stop it verifying the TLS cert. However, this causes an
InsecureRequestWarning error in the home assistant log.

Removing this flag seems to work okay, at least with Python 3.13.3, and
it's bad practice not to verify TLS certs when making HTTP requests, so
remove the flag.